Frontend - Move event from from events to its proper page
